requirements.txt
pandas
progressbar2
python-utils
tensorflow == 1.12.0rc2
numpy
matplotlib
scipy
sox
paramiko >= 2.1
python_speech_features
pyxdg
bs4
six
requests
tables
attrdict
